// frontend/src/generated-api/models/GetProjectDetailResponseDto.ts
export interface GetProjectDetailResponseDto {
    /**
     * 프로젝트 id
     * @type {number}
     * @memberof GetProjectDetailResponseDto
     */
    projectId?: number;
    /**
     * 프로젝트명
     * @type {string}
     * @memberof GetProjectDetailResponseDto
     */
    title?: string;
    /**
     * 프로젝트 설명
     * @type {string}
     * @memberof GetProjectDetailResponseDto
     */
    description?: string;
    /**
     * 모델 타입
     * @type {string}
     * @memberof GetProjectDetailResponseDto
     */
    modelType?: GetProjectDetailResponseDtoModelTypeEnum;
    /**
     * 모델명
     * @type {string}
     * @memberof GetProjectDetailResponseDto
     */
    modelName?: string;
    /**
     * 프로젝트 생성일
     * @type {string}
     * @memberof GetProjectDetailResponseDto
     */
    createdAt?: string;
    /**
     * 프로젝트 수정일
     * @type {string}
     * @memberof GetProjectDetailResponseDto
     */
    updatedAt?: string;
    /**
     * 
     * @type {SlideSummaryDto}
     * @memberof GetProjectDetailResponseDto
     */
    slideSummary?: SlideSummaryDto;
    /**
     * 
     * @type {ModelProcessDto}
     * @memberof GetProjectDetailResponseDto
     */
    modelProcess?: ModelProcessDto;
    /**
     * 
     * @type {Array<LabelDto>}
     * @memberof GetProjectDetailResponseDto
     */
    labels?: Array<LabelDto>;
    /**
     * 
     * @type {Array<SlideDto>}
     * @memberof GetProjectDetailResponseDto
     */
    slides?: Array<SlideDto>;
    /**
     * 
     * @type {AnalyticsDto}
     * @memberof GetProjectDetailResponseDto
     */
    analytics?: AnalyticsDto;
}


/**
 * @export
 */
export const GetProjectDetailResponseDtoModelTypeEnum = {
    Tissue: 'TISSUE',
    Cell: 'CELL',
    Multi: 'MULTI'
} as const;
export type GetProjectDetailResponseDtoModelTypeEnum = typeof GetProjectDetailResponseDtoModelTypeEnum[keyof typeof GetProjectDetailResponseDtoModelTypeEnum];